Ensure latest version is used, not latest release.
Realized there was an interesting condition that I was not handling.
The github releases api sorts the releases by the date they were
created, but that doesn't necessarily mean that it's the latest version
(e.g. a new version of an older branch is released).

The previous way of just taking the latest tag and downloading a release
for it doesn't work either because the assets may not be built yet for
that tag.

This change simply walks backward through the latest tags until a
release is found, and uses that.